Finalised implementation of Model::growAtom(). Added 'growAtom' command. Added complementary nBonds data for Atom::AtomGeometry. Renamed UnitCell::mimd() to UnitCell::mimVector(), and made argument ordering more consistent (now returns mim vector of r1 to r2). Removed old code from Atom::nextBondGeometry() and deleted associated old routines, replacing with new generic code for any geometry (nearly complete). Renamed CellVariable 'mimd' to 'mimVector'. Manual update.

+++ src/model/model.h
@@ -711,7 +711,7 @@ class Model
 
 	public:
 	// Add a single atom of the type specified to the atom specified
-	Atom *growAtom(Atom *target, int element, double distance, Atom::AtomGeometry geometry, bool bound = TRUE);
+	Atom *growAtom(Atom *target, short int element, double distance, Atom::AtomGeometry geometry, bool bound = TRUE);
 	// Adds hydrogens to satisfy the bond order requirements of atoms in the model
 	void hydrogenSatisfy(Atom *target = NULL);
 	// Return the pen orientation matrix
